+++ R/terra-connectivity.R
@@ -335,6 +335,7 @@ aggregate_connected_patches <- function(raster) {
 #'
 #' @param habitat Terra SpatRaster. Habitat raster.
 #' @param barrier Terra SpatRaster. Barrier raster.
+#' @param species Species name. E.g., "Blue-tongued Lizard".
 #' @param interpatch_distance Numeric. The distance (in meters) where habitat
 #'   patches are considered connected. E.g., if set to 500, patches 498m apart
 #'   are connected, those 501m apart are not connected. This is passed
@@ -366,12 +367,14 @@ aggregate_connected_patches <- function(raster) {
 #' connectivity <- habitat_connectivity(
 #'     habitat = lizard_habitat,
 #'     barrier = lizard_barrier,
-#'     interpatch_distance = 10
+#'     species = "Blue-tongued Lizard",
+#'     interpatch_distance = 12
 #'   )
 #' connectivity
 habitat_connectivity <- function(
   habitat,
   barrier,
+  species,
   interpatch_distance = NULL,
   buffer_radius = NULL,
   verbose = TRUE
@@ -393,6 +396,13 @@ habitat_connectivity <- function(
       buffer_radius
     )$result
   }
+  habitat_connectivity <- patch_connectivity(
+    data = habitat_connectivity,
+    species = species,
+    # store the FULL distance
+    interpatch_distance = buffer_radius * 2
+  )
+
   habitat_connectivity
 }