+++ data-raw/lizard-connected-example.R
@@ -2,8 +2,9 @@ library(urbioconnect)
 lizard_areas_connected <- habitat_connectivity(
   habitat = example_habitat(),
   barrier = example_barrier(),
+  species = "Blue-tongued Lizard",
   interpatch_distance = 50,
   verbose = FALSE
 )
 
-usethis::use_data(lizard_areas_connected)
+usethis::use_data(lizard_areas_connected, overwrite = TRUE)